Allow config string in `run_and_get_multiple_output` and correct problem in call with 'box'
This commit alters the construction of the config string to allow additional parameters. It also ensures 'box' can be listed before other extensions: as 'box' adds configfiles to the config string, later tessedit parameters (specifically for 'tsv' and 'xml') must be preceded by '-c', or are otherwise silently ignored, resulting in no output for those extensions.